Handle remainder properly
Do an integer division on the number of players vs the table capacity
to get the number of total tables. Multiply by 4 then subtract from the
full number of players. The subtraction results in the number of bots
that needs to be added to get to a clean multiple of 4.